Continued recaftoring of js code with omkar.
