+++ godot-src/gameOrchestrator.gd
@@ -1,0 +1,17 @@
+        speechRecognizer.startListeningStop listening
+        Set dialogue state
+
+
+        Prompt handler call
+        start conversation
+        response = add turn()
+
+        // you have text
+
+        await (generateViseme(text), generateAudio(text))
+
+
+        generateSpeech(window.aiResponse);
+            speechsyntesizer (playViseme())
+
+
